cleanup, aligning with Nemesys and separating some code from speedTester. Reintroduced upload in mist.

+++ mist/log_conf.py
@@ -17,8 +17,9 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import logging.config
+import os
+
 import paths
-#import re
 
 configfile = paths.CONF_LOG
 logfile = paths.LOG_FILE
@@ -85,16 +86,10 @@ format=%(asctime)s MIST %(filename)s.%(funcName)s():%(lineno)d [%(levelname)s] %
 datefmt=%b %d %H:%M:%S
 '''
 
-def init_log():
-    with open(configfile, 'w') as f:
-        'TODO: choose which one to use'
-        s = str(default_no_stdout)
-        f.write(s) 
+def init_log(level=logging.INFO, use_name='MIST'):
+    if not os.path.isfile(configfile):
+        with open(configfile, 'w') as f:
+            s = str(default_no_stdout)
+            f.write(s) 
 
     logging.config.fileConfig(configfile, disable_existing_loggers=False)
-    
-# create logger
-# class Logger(logging.getLoggerClass()):
-# 
-#     def __init__(self):
-#         init_log()
